Improved code abstraction, removed unnecessary code, modified packages name
