+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.PatientAppCompose"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.pages.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.PatientAppCompose">
