+++ app/src/main/java/com/example/patientappcompose/data/di/NetworkModule.kt
@@ -12,7 +12,6 @@ import dagger.hilt.components.SingletonComponent
 import retrofit2.Converter.Factory
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-import retrofit2.create
 import javax.inject.Singleton
