+++ app/src/main/java/com/example/patientappcompose/ui/composable/components/CustomRatingBar.kt
@@ -1,4 +1,4 @@
-package com.example.patientappcompose.ui.composable
+package com.example.patientappcompose.ui.composable.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.height
