+++ app/src/main/java/com/example/patientappcompose/ui/composable/components/PatientList.kt
@@ -1,4 +1,4 @@
-package com.example.patientappcompose.ui.composable
+package com.example.patientappcompose.ui.composable.components
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
@@ -9,22 +9,18 @@ import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import com.example.patientappcompose.domain.model.patients.PatientDataModel
-import com.example.patientappcompose.ui.features.details.DetailsViewModel
-import com.example.patientappcompose.ui.features.patients.PatientsViewModel
+import com.example.patientappcompose.ui.composable.containers.PatientRow
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun PatientList(
     data: List<PatientDataModel>,
-    patientViewModel: PatientsViewModel,
-    detailsViewModel: DetailsViewModel,
+    patientId: String?,
     onClick: (PatientDataModel) -> Unit,
     onDelete: (PatientDataModel) -> Unit
 ) {
-    val context = LocalContext.current
     LazyVerticalGrid(
         columns = GridCells.Fixed(2),
         modifier = Modifier
@@ -43,8 +39,9 @@ fun PatientList(
                 PatientRow(
                     record = data[index],
                     onClick = { onClick(data[index]) },
-                    selectedPatientId = patientViewModel.selectedPatientId,
-                ) { onDelete(data[index]) }
+                    selectedPatientId = patientId,
+                    onDelete = { onDelete(data[index]) }
+                )
             }
         }
     }
