+++ app/src/main/java/com/example/patientappcompose/ui/navigator/Navigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.patientappcompose.ui.screens.AddPatientScreen
-import com.example.patientappcompose.ui.screens.PatientDetailScreen
+import com.example.patientappcompose.ui.screens.DetailPatientScreen
 import com.example.patientappcompose.ui.screens.PatientsScreen
 
 @Composable
@@ -21,7 +21,7 @@ fun Navigation( ) {
         composable(route = Screens.PatientDetailsScreen.route){backStackEntry ->
             // You can retrieve the patientId like this
             val patientId = backStackEntry.arguments?.getString("patientId")
-            PatientDetailScreen(navController, patientId ?: "")
+            DetailPatientScreen(navController, patientId ?: "")
         }
     }
 }
