+++ app/src/main/java/com/example/patientappcompose/ui/pages/MainActivity.kt
@@ -1,16 +1,8 @@
-package com.example.patientappcompose.ui
+package com.example.patientappcompose.ui.pages
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.patientappcompose.ui.navigator.Navigation
 import com.example.patientappcompose.ui.theme.PatientAppComposeTheme
 import dagger.hilt.android.AndroidEntryPoint
