+++ app/src/main/java/com/example/patientappcompose/ui/screens/PatientsScreen.kt
@@ -4,11 +4,13 @@ package com.example.patientappcompose.ui.screens
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.example.patientappcompose.ui.composable.AllPatientsLogic
-import com.example.patientappcompose.ui.composable.CircularBar
-import com.example.patientappcompose.ui.composable.FloatButton
+import com.example.patientappcompose.ui.composable.logic.AllPatientsLogic
+import com.example.patientappcompose.ui.composable.components.CircularBar
+import com.example.patientappcompose.ui.composable.components.FloatButton
+import com.example.patientappcompose.ui.composable.containers.ErrorMessage
 import com.example.patientappcompose.ui.features.patients.PatientsViewModel
 import com.example.patientappcompose.ui.navigator.Screens
 
@@ -18,9 +20,9 @@ fun PatientsScreen(
     navController: NavController,
     viewModel:PatientsViewModel = hiltViewModel()
 ) {
-    val data = viewModel.patientStateFlow.collectAsState().value
-    val loading = viewModel.loadingStateFlow.collectAsState().value
-    val error = viewModel.errorStateFlow.collectAsState().value
+    val data by viewModel.patientStateFlow.collectAsState()
+    val loading by viewModel.loadingStateFlow.collectAsState()
+    val error by viewModel.errorStateFlow.collectAsState()
 
 
 
@@ -30,9 +32,7 @@ fun PatientsScreen(
     if (loading){
         CircularBar()
     }
-    if (error != null){
-        ErrorMessage(error = error)
-    }
+    error?.let { ErrorMessage(error = it) }
 
     FloatButton {
         navController.navigate(Screens.AddScreen.route)
